Fix comparison between signed and unsigned integers

+++ src/graphics/particle/ParticleEffects.h
@@ -131,7 +131,7 @@ struct PARTICLE_DEF {
 
 //-----------------------------------------------------------------------------
 
-static const int MAX_POLYBOOM = 4000;
+static const size_t MAX_POLYBOOM = 4000;
 static const float FLARE_MUL = 2.f;
 
 static const float BOOM_RADIUS = 420.f;
